refactor(cli): split server_logs.go and consolidate search display rendering
Split server_logs.go (442 lines) into server_logs.go (244, command+config)
and server_logs_stream.go (200, streaming mechanics) to comply with the
250-line guideline.

Add generic proto slice rendering functions (RenderProtoSliceJSON[T],
RenderProtoSliceYAML[T], DisplayProtoSlice[T]) to pkg/display/proto.go
and refactor search/display.go to use them, eliminating duplicated
protojson.MarshalOptions and fixing broken JSON array indentation.

- server_logs.go: mechanical 2-file split, zero logic changes
- pkg/display/proto.go: +3 generic functions, +encoding/json import
- pkg/display/proto_test.go: +8 tests for slice rendering
- search/display.go: 300→233 lines, removed displayResultsYAML/JSON
- search/BUILD.bazel: removed protojson and yaml.v3 deps

Made-with: Cursor
