+++ _projects/2026-02/20260226.01.cli-output-system-refactor/checkpoints/2026-02-26-session-9.md
@@ -1,0 +1,34 @@
+# Session Notes: 2026-02-26 Session 9
+
+## Accomplishments
+
+- **Item 5**: Split `server_logs.go` (442 lines) into `server_logs.go` (244) + `server_logs_stream.go` (200). Pure mechanical split, both under 250-line guideline.
+- **Item 7**: Added generic proto slice rendering functions (`RenderProtoSliceJSON[T]`, `RenderProtoSliceYAML[T]`, `DisplayProtoSlice[T]`) to `pkg/display/proto.go`. Refactored `search/display.go` from 300 to 233 lines by replacing hand-built YAML/JSON rendering with shared utilities. Fixed broken JSON array indentation. 8 new tests.
+
+## Decisions Made
+
+- **2-file split over 3-file**: Config helpers (~70 lines) stay with command definition because they're only called from `Run`. A separate `server_logs_config.go` would have been too thin.
+- **Go generics for slice functions**: Using `[T proto.Message]` avoids forcing callers to convert concrete slices. Precedent in `synthesis/reader.go`.
+- **`json.MarshalIndent` over manual array building**: Fixes broken indentation, simpler code. Behavioral change is an improvement.
+- **Batch error over per-entry continue**: Simpler. Proto marshaling of valid gRPC responses is unreachable per Phase 4 DD#16.
+
+## Key Code Changes
+
+- `server_logs.go`: Trimmed from 442 to 244 lines (streaming functions extracted)
+- `server_logs_stream.go`: New file with 5 streaming functions (200 lines)
+- `pkg/display/proto.go`: +68 lines (3 generic slice functions using `encoding/json` for proper array formatting)
+- `pkg/display/proto_test.go`: +98 lines (8 new tests: JSON/YAML/empty/dispatcher for slices)
+- `search/display.go`: -67 lines (removed `displayResultsYAML`, `displayResultsJSON`, direct proto/yaml imports)
+- `search/BUILD.bazel`: Removed `protojson` and `yaml.v3` deps
+
+## Learnings
+
+- The manual JSON array builder in `search/display.go` (`fmt.Println("[")` + comma logic) produced broken indentation — only the first line of each entry was indented within the array. `json.MarshalIndent` over `[]json.RawMessage` is the correct approach for proto-to-JSON arrays.
+
+## Open Questions
+
+- None for this session.
+
+## Next Session Plan
+
+- Remaining deferred items: Item 3 (`cliprint` sunset), Item 4 (table rendering modernization), Item 6 (integration tests for `--json`/`--quiet`), ProgressDisplay migration, TUI icon vocabulary consistency.
