+++ client-apps/cli/cmd/stigmer/root/BUILD.bazel
@@ -53,6 +53,7 @@ go_library(
         "server_health.go",
         "server_llm.go",
         "server_logs.go",
+        "server_logs_stream.go",
         "server_status.go",
         "validate.go",
         "verb_helpers.go",
