+++ client-apps/cli/internal/cli/search/BUILD.bazel
@@ -17,9 +17,7 @@ go_library(
         "//client-apps/cli/pkg/display",
         "@com_github_fatih_color//:color",
         "@com_github_pkg_errors//:errors",
-        "@in_gopkg_yaml_v3//:yaml_v3",
         "@org_golang_google_grpc//:grpc",
-        "@org_golang_google_protobuf//encoding/protojson",
     ],
 )
