+++ client-apps/cli/pkg/display/proto.go
@@ -1,6 +1,7 @@
 package display
 
 import (
+	"encoding/json"
 	"fmt"
 	"io"
 	"os"
@@ -66,6 +67,73 @@ func DisplayProto(msg proto.Message, format string, tableFunc func()) {
 		}
 	case "json":
 		if err := RenderProtoJSON(os.Stdout, msg); err != nil {
+			fmt.Fprintf(os.Stderr, "Error: %v\n", err)
+		}
+	default:
+		tableFunc()
+	}
+}
+
+// RenderProtoSliceJSON marshals a slice of proto messages as a properly
+// indented JSON array and writes it to w. Each element is marshaled with the
+// shared protoMarshalOptions, then the array is formatted via encoding/json
+// to produce correct nested indentation.
+func RenderProtoSliceJSON[T proto.Message](w io.Writer, items []T) error {
+	raw := make([]json.RawMessage, len(items))
+	for i, item := range items {
+		b, err := protoMarshalOptions.Marshal(item)
+		if err != nil {
+			return fmt.Errorf("failed to marshal item %d to JSON: %w", i, err)
+		}
+		raw[i] = json.RawMessage(b)
+	}
+
+	out, err := json.MarshalIndent(raw, "", "  ")
+	if err != nil {
+		return fmt.Errorf("failed to marshal JSON array: %w", err)
+	}
+	_, err = fmt.Fprintln(w, string(out))
+	return err
+}
+
+// RenderProtoSliceYAML marshals a slice of proto messages as a YAML array and
+// writes it to w. Each element is round-tripped through JSON (via protojson)
+// to respect proto field naming conventions, then the whole slice is marshaled
+// as YAML.
+func RenderProtoSliceYAML[T proto.Message](w io.Writer, items []T) error {
+	entries := make([]map[string]interface{}, len(items))
+	for i, item := range items {
+		b, err := protoMarshalOptions.Marshal(item)
+		if err != nil {
+			return fmt.Errorf("failed to marshal item %d to JSON: %w", i, err)
+		}
+		var m map[string]interface{}
+		if err := yaml.Unmarshal(b, &m); err != nil {
+			return fmt.Errorf("failed to parse item %d for YAML conversion: %w", i, err)
+		}
+		entries[i] = m
+	}
+
+	yamlBytes, err := yaml.Marshal(entries)
+	if err != nil {
+		return fmt.Errorf("failed to marshal to YAML: %w", err)
+	}
+	_, err = fmt.Fprint(w, string(yamlBytes))
+	return err
+}
+
+// DisplayProtoSlice is the array counterpart of DisplayProto. For "yaml" and
+// "json" formats it renders the proto slice to stdout using the shared
+// marshaling utilities. For any other format (including "table") it delegates
+// to tableFunc.
+func DisplayProtoSlice[T proto.Message](items []T, format string, tableFunc func()) {
+	switch format {
+	case "yaml":
+		if err := RenderProtoSliceYAML(os.Stdout, items); err != nil {
+			fmt.Fprintf(os.Stderr, "Error: %v\n", err)
+		}
+	case "json":
+		if err := RenderProtoSliceJSON(os.Stdout, items); err != nil {
 			fmt.Fprintf(os.Stderr, "Error: %v\n", err)
 		}
 	default:
